gst: set the unset flag on Buffers fields
These buffer fields are unset so set the flag accordingly.
In 0.2 this didn't really negotiate to a common set of parameters
but that was then ignored and a default set of params was used
and things worked.

In the compat layer, however, this causes problems because a failed
negotiation will now result in an error.

Setting the unset flag is the right thing to do but we might have
to do a workaround on the server to support the old broken definition.